+++ lib/amp_sdk/command_runner.ex
@@ -1,16 +1,14 @@
 defmodule AmpSdk.CommandRunner do
   @moduledoc false
 
-  alias AmpSdk.{Command, Error}
-
-  @default_timeout_ms 60_000
+  alias AmpSdk.{Command, Defaults, Error}
 
   @type run_result :: {:ok, String.t()} | {:error, Error.t()}
 
   @spec run([String.t()], keyword()) :: run_result()
   def run(args, opts \\ [])
       when is_list(args) and is_list(opts) do
-    run(args, opts, @default_timeout_ms)
+    run(args, opts, Defaults.command_timeout_ms())
   end
 
   @spec run([String.t()], keyword(), non_neg_integer()) :: run_result()
@@ -22,5 +20,5 @@ defmodule AmpSdk.CommandRunner do
   end
 
   @spec default_timeout_ms() :: pos_integer()
-  def default_timeout_ms, do: @default_timeout_ms
+  def default_timeout_ms, do: Defaults.command_timeout_ms()
 end